+++ com.unity.visualeffectgraph/Editor/Models/Operators/Implementations/LogicalNot.cs
@@ -11,12 +11,13 @@ namespace UnityEditor.VFX.Operator
         public class InputProperties
         {
             static public bool FallbackValue = false;
-            [Tooltip("The operand.")]
+            [Tooltip("Sets the operand.")]
             public bool a = FallbackValue;
         }
 
         public class OutputProperties
         {
+            [Tooltip("Outputs the opposite of the operand.")]
             public bool o = false;
         }