+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -107,6 +107,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed an issue with fog volume component upgrade not working properly
 - Fixed Spot light Pyramid Shape has shadow artifacts on aspect ratio values lower than 1
 - Fixed issue with AO upsampling in XR
+- Fixed camera without HDAdditionalCameraData component not rendering
+- Removed the macro ENABLE_RAYTRACING for most of the ray tracing code
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled